requirements.txt
# Core dependencies
playwright==1.41.0
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-core==0.1.10
langgraph==0.0.20

# Async and HTTP
aiohttp==3.9.1
httpx==0.26.0

# Data validation and parsing
pydantic==2.5.3
pydantic-settings==2.1.0

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# Utilities
python-dateutil==2.8.2
pillow==10.2.0

# CLI
click==8.1.7
rich==13.7.0

# Logging
loguru==0.7.2

# Google Gemini
google-generativeai>=0.4.1

